requirements.txt
lxml==5.3.0
pymongo==4.6.3
PyMySQL==1.1.1
Requests==2.32.0
schedule==1.2.1
tqdm==4.66.3
Pillow>=9.0.0
pytesseract>=0.3.8
